+++ Util/SimRobot/Src/SimRobotCore2/Tools/OpenGLTools.cpp
@@ -10,7 +10,7 @@
 
 void OpenGLTools::computePerspective(float fovy, float aspect, float near, float far, float matrix[])
 {
-  matrix[5] = 1.f / tanf(fovy * 0.5f);
+  matrix[5] = 1.f / std::tan(fovy * 0.5f);
   matrix[0] = matrix[5] / aspect;
   const float nearMFarInv = 1.f / (near - far);
   matrix[10] = (far + near) * nearMFarInv;
